requirements.txt
pandas
openpyxl
ortools
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0